requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
sqlalchemy_utils
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
psycopg2